+++ CHANGELOG.md
@@ -5,6 +5,10 @@ Changelog
 
 * Updates the bugsnag-android dependency from v5.28.1 to [v5.28.3](https://github.com/bugsnag/bugsnag-android/blob/master/CHANGELOG.md#5283-2022-11-16)
 
+## TBD
+
+* Updates the bugsnag-cocoa dependency from v6.25.0 to [v6.25.2](https://github.com/bugsnag/bugsnag-cocoa/blob/master/CHANGELOG.md#6252-2023-01-18)
+
 ## 1.6.0 (2022-11-09)
 
 * Adds the `Telemetry.Usage` configuration property. [#194](https://github.com/bugsnag/bugsnag-unreal/pull/194)