+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Changelog
 
 * Adds the `Telemetry.Usage` configuration property. [#194](https://github.com/bugsnag/bugsnag-unreal/pull/194)
 * Adds the `MaxStringValueLength` configuration property. [#193](https://github.com/bugsnag/bugsnag-unreal/pull/193)
+* Updates the package branding
 
 ## 1.5.1 (2022-10-27)